Fix migration script. Fix Type, Reminder, ReminderDataProvider, JsonModel classes.

+++ app/src/main/java/com/cray/software/justreminder/datas/models/ShoppingList.java
@@ -150,7 +150,7 @@ public class ShoppingList {
         this.id = id;
     }
 
-    public long getStatus() {
+    public int getStatus() {
         return status;
     }
